requirements.txt
hubspot-api-client==7.5.0
pprintpp==0.4.0
DateTime==5.1
zope.interface==6.0
python-crontab==2.7.1